+++ src/ui/buttonDropdown/types.ts
@@ -1,0 +1,7 @@
+import { IDropdownPublicProps } from './../dropdown/types';
+
+export interface IButtonDropdownProps extends
+    IDropdownPublicProps
+{
+    disabled?: boolean;
+}
